backend/requirements.txt
# Kyotech AI Backend - Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.19

azure-storage-blob>=12.24.0
openai>=1.59.2

asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
pgvector>=0.3.6

pymupdf>=1.25.0

python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0

PyJWT[crypto]>=2.8.0
httpx>=0.27.0

scalar-fastapi>=1.0.0